+++ hnetw/ui/qtmodels/hn_packet_list_row.h
@@ -5,7 +5,7 @@
 class HnPacketListRow {
 public:
     HnPacketListRow();
-    HnPacketListRow(HnPacket* packet, long packetOffset);
+    HnPacketListRow(HnPacket* packet, size_t packetOffset);
 
     ~HnPacketListRow();
 
@@ -19,8 +19,8 @@ public:
     };
 
     const QString columnString(int column);
-    long packetOffset() const;
-    int packetLength() const;
+    size_t packetOffset() const;
+    size_t packetLength() const;
 
 private:
     void setId(int id);
@@ -28,11 +28,11 @@ private:
     void setSourceIp(uint32_t srcIp);
     void setDestinationIp(uint32_t destIp);
     void setProtocolType(QString type);
-    void setLength(int length);
-    void setCaptureFileOffset(long packetOffset, int packetLen);
+    void setLength(size_t length);
+    void setCaptureFileOffset(size_t packetOffset, size_t packetLen);
 
-    long packetOffset_ = 0;
-    int packetLen_ = 0;
+    size_t packetOffset_ = 0;
+    size_t packetLen_ = 0;
     const static size_t columnsNum_ = 6;
     std::array<QString, columnsNum_> rowColumns_ = {};
 };
